backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# HTTP & WebSocket
httpx==0.28.1
websockets==14.1

# AI & Services
openai==1.58.1
google-generativeai==0.8.3
google-cloud-speech==2.24.1
elevenlabs==1.50.3

# Rate Limiting
slowapi==0.1.9

# Redis
redis==5.2.1

# CORS
python-multipart==0.0.17

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
coverage==7.6.9
